requirements.txt
numpy
pandas
seaborn
python-dotenv
#-e .